error: Move definition of telex_error to internal header
Because the definition of `struct telex_error` is in error.c, other
parts of the library cannot access members of the structure directly.

This commit moves the definition of `struct telex_error` to an
internal header, so that other parts of the library may directly
access members of the structure.