token: Add operators for more intuitive expressions
The '+' and '-' operators are not very intuitive because they do not
graphically reflect the direction that they represent. The language
also needs operators for jumping to the end of a match (instead of
the beginning, like '+' and '-'). For this purpose, the following
operators are more intuitive.

 - '<', '>': jump to the beginning of a match
 - '<<', '>>': jump to the end of a match

This commit removes the '+' and '-' operators from the lexer and
adds the '>', '>>', '<', '<<' operators instead. This commit further
adds the '#' operator, which may optionally be used to denote
movement in a line (as counterpart to ':').