token: Make next_relevant() available to other parts of the library
The `next_relevant()` function needs to be used by other parts of the
library to implement the telex parser, but it is a static function.

This commit removes the static specifier from `next_relevant()`,
making it callable from other parts of the library. This commit also
renames the function to `next_relevant_token()` to make its purpose
more obvious, when used outside of token.c.