token: Make `struct token` available to other parts of the library
The parser needs access to members of `struct token` but it is an
opaque data type. Because there are no getter/setter functions for
all members, not all members can be accessed.

This commit moves the definition of `struct token` into a header
that is shared within the library, allowing other parts of the
library to access its members directly.