parser: Implement parsing of nested telexes
The parser does not handle parentheses, making it impossible to
nest telexes, which greatly reduces their utility.

This commit modifies the parser and its grammar so that expressions
can be nested. This change allows the `|` operator to be used
anywhere in an expression, not just between string and regex matches.