token: Make token_to_string() behave like snprintf()
To implement `telex_to_string()`, a method for writing tokens to a
string with a behavior similar to `snprintf()` is necessary. The
token type has a `token_to_string()` method, which allows access to
a token's lexeme, but it isn't necessary because token is an
internal type, and its members can be directly accessed. Thus, the
function should be repurposed to behave like `snprintf()`.

This commit modifies the `token_to_string()` function so it writes
the lexeme of a token to a buffer with a user-specified size.