parser: Fix use of variable arguments to get_token()
Because `get_token()` uses variable argument lists, it expects the
last argument after a list of token types to be `0`, but the parser
does not always pass `0` after the last token type, which may
trigger bugs if invalid inputs are parsed.

This commit modifies the parser so that it always passes `0` as the
last argument to `get_token()`.