telex: Add method to distinguish relative from absolute telexes
When handling telexes, applications may need to distinguish between
relative and absolute telexes, but there is no way for them to find
out if a telex is one or the other.

This commit adds the `telex_is_relative()` method, which determines
if a telex is relative by looking at its prefix.