eval: Make '<<'/'>>' operators skip to the start/end of a string
The `<<` and `>>` operators are supposed to return a pointer to
the beginning and end of a string, respectively. However, neither
operator behaves correctly.

This commit modifies the `eval_string()` function so that it adds
or subtracts the length of the match if the search string was
prefixed with either `<<` or `>>`.