eval: Correctly evaluate relative line expressions
Relative line expressions are not evaluated correctly, leading to
incorrect pointers to be returned by `telex_lookup()`.

This commit rewrites the `eval_line_expr()` function to make sure
relative line expressions are evaluated correctly. This also fixes
the operator behavior to return pointers to the line beginnings or
line endings, depending on the search direction.