eval: Make absolute line expressions refer to the correct lines
The absolute expression `:1' should refer to the first line of the
file, but the evaluation function returns a pointer to the second
line, which is intuitively incorrect. When using absolute telexes,
the first line should be `:1', not `:0'.

This commit modifies the `eval_line_expr()' method so that it
reduces the number of steps by one when evaluating an absolute
expression.