error: Add methods for freeing telex_error types
There are no functions for freeing telex_error types. Thus, users
are tempted to use `free()' instead, which will cause memory leaks
because memory that is associated with telex_errors is not freed.

This commit adds the methods `telex_error_free()' and
`telex_error_free_all()', which can be used to free one or multiple
telex_error types, respectively.