telex: Add function to determine telex from position in a buffer
There are use-cases where it may be beneficial to be able to perform
a reverse lookup of a telex. That is, a user might want to determine
a telex that refers to a certain position in a buffer.

This commit adds the `telex_rlookup()' method which performs a
reverse lookup, determining the telex for a position in a buffer.