eval: Return pointer to end of last line when seeking behind buffer
When evaluating a telex that seeks to a line behind the end of the
buffer, the evaluator returns a pointer to the start of the last
line. However, it would be more intuitive to seek to the end of the
last line in this case.

This commit modifies `eval_line_expr()' so that it returns a pointer
to the end of the last line if the evaluated expression seeks to a
line behind the end of the buffer.