telex: Implement concatenation of multiple telexes
The `telex_combine()' method is not implemented, thus preventing
applications from concatenating telexes.

This commit modifies `telex_combine()' so that it combines (copies
of) the left and right telexes by combining their parse trees as
sub-trees of one combined tree.
The parameters of `telex_combine()' have been made `const' to
emphasize that the original telexes will be left untouched.