telex: Clone tokens when cloning expressions that refer to them
When cloning telexes (or parts of telexes), the tokens that the
expressions refer to are not cloned. Thus, when freeing a telex
that was cloned, the tokens of both telexes will be freed, causing
use-after-free bugs.

This commit modifies all clone methods so that they also clone
tokens that the expressions refer to.